refactor: create a class helper to create tree
