--- android/beagle/src/main/java/br/com/zup/beagle/android/jsonpath/JsonPathReplacer.kt
@@ -1,34 +1,0 @@
-/*
- * Copyright 2020 ZUP IT SERVICOS EM TECNOLOGIA E INOVACAO SA
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package br.com.zup.beagle.android.jsonpath
-
-import java.util.*
-
-internal class JsonPathReplacer(
-    private val jsonCreateTree: JsonCreateTree = JsonCreateTree()
-) {
-
-    fun replace(keys: LinkedList<String>, newValue: Any, root: Any): Pair<Boolean, Any> {
-        return when {
-            keys.isEmpty() -> false to root
-            else -> {
-                val newRoot = jsonCreateTree.walkingTreeAndFindKey(root, keys, newValue)
-                return true to newRoot
-            }
-        }
-    }
-}
